requirements.txt
fastapi~=0.88.0
pycodestyle
pydantic~=1.10.2
SQLAlchemy~=1.4.45
starlette~=0.22.0
toml
uvicorn~=0.20.0
psycopg2-binary
passlib~=1.7.4
bcrypt
python-multipart
python-jose
cryptography
postgres



